+++ CHANGELOG.md
@@ -4,6 +4,11 @@ This project uses [Semantic Versioning](https://semver.org/) - MAJOR.MINOR.PATCH
 
 # Changelog
 
+## 1.0.1 (2026-03-14)
+
+- Added configuration handling for the built in lighthouse DNS server
+- Added configuration handling for the built in lighthouse SSHD server
+
 ## 1.0.0 (2026-03-05)
 
 - Initial release of saltext-nebula